requirements.txt
Flask>=2.0
google-api-python-client>=2.70.0
google-auth>=2.20.0
google-auth-httplib2>=0.1.0
werkzeug>=2.0
python-magic==0.4.27
Pillow>=9.0.0
